==> scripts/00a_renv-set-up.R <==
## 
## R environment set-up
##
## -------------------------------------------------------------------------- ##
## -------------------------------------------------------------------------- ##

# renv::init()
# renv::snapshot()

## If you don't have {renv} installed, please uncomment 
## the line below and install the package

# install.packages("renv")

## Then, just run the following command, which will restore the R environment 
## using the same package versions as we used

renv::restore()

## If you're not familiarised with {renv}, please have a look at:
## https://rstudio.github.io/renv/articles/renv.html
